@@ -1,12 +1,11 @@
 class Token {
-  BankAccount bankAccount;
-  CreditCard card;
-  double created;
-  bool livemode;
-  String tokenId;
-  Extra extra;
+  BankAccount? bankAccount;
+  CreditCard? card;
+  double? created;
+  bool? livemode;
+  String? tokenId;
 
-  Token({this.bankAccount, this.card, this.created, this.livemode, this.tokenId, this.extra});
+  Token({this.bankAccount, this.card, this.created, this.livemode, this.tokenId});
 
   factory Token.fromJson(Map<dynamic, dynamic> json) {
     return Token(
@@ -15,7 +14,6 @@ class Token {
       created: json['created'] is int ? (json['created'] as int).toDouble() : json['created'],
       livemode: json['livemode'],
       tokenId: json['tokenId'],
-      extra: json['extra'] != null ? Extra.fromJson(json['extra']) : null,
     );
   }
 
@@ -25,28 +23,25 @@ class Token {
     if (this.livemode != null) data['livemode'] = this.livemode;
     if (this.tokenId != null) data['tokenId'] = this.tokenId;
     if (this.bankAccount != null) {
-      data['bankAccount'] = this.bankAccount.toJson();
+      data['bankAccount'] = this.bankAccount!.toJson();
     }
     if (this.card != null) {
-      data['card'] = this.card.toJson();
-    }
-    if (this.extra != null) {
-      data['extra'] = this.extra.toJson();
+      data['card'] = this.card!.toJson();
     }
     return data;
   }
 }
 
 class BankAccount {
-  String accountHolderName;
-  String accountHolderType;
-  String accountNumber;
-  String bankName;
-  String countryCode;
-  String currency;
-  String fingerprint;
-  String last4;
-  String routingNumber;
+  String? accountHolderName;
+  String? accountHolderType;
+  String? accountNumber;
+  String? bankName;
+  String? countryCode;
+  String? currency;
+  String? fingerprint;
+  String? last4;
+  String? routingNumber;
 
   BankAccount(
       {this.accountHolderName,
@@ -89,23 +84,24 @@ class BankAccount {
 }
 
 class CreditCard {
-  String addressCity;
-  String addressCountry;
-  String addressLine1;
-  String addressLine2;
-  String addressState;
-  String addressZip;
-  String brand;
-  String cardId;
-  String country;
-  int expMonth;
-  int expYear;
-  String funding;
-  String last4;
-  String name;
-  String number;
-  String cvc;
-  String token;
+  String? addressCity;
+  String? addressCountry;
+  String? addressLine1;
+  String? addressLine2;
+  String? addressState;
+  String? addressZip;
+  String? brand;
+  String? cardId;
+  String? country;
+  String? currency;
+  int? expMonth;
+  int? expYear;
+  String? funding;
+  String? last4;
+  String? name;
+  String? number;
+  String? cvc;
+  String? token;
 
   CreditCard(
       {this.addressCity,
@@ -116,6 +112,7 @@ class CreditCard {
       this.addressZip,
       this.brand,
       this.cardId,
+      this.currency,
       this.country,
       this.expMonth,
       this.expYear,
@@ -136,6 +133,7 @@ class CreditCard {
         addressZip: json['addressZip'],
         brand: json['brand'],
         cardId: json['cardId'],
+        currency: json['currency'],
         country: json['country'],
         expMonth: json['expMonth'],
         expYear: json['expYear'],
@@ -161,99 +159,12 @@ class CreditCard {
     if (this.expMonth != null) data['expMonth'] = this.expMonth;
     if (this.expYear != null) data['expYear'] = this.expYear;
     if (this.funding != null) data['funding'] = this.funding;
+    if (this.currency != null) data['currency'] = this.currency;
     data['last4'] = this.last4;
     if (this.name != null) data['name'] = this.name;
     if (this.number != null) data['number'] = this.number;
     if (this.cvc != null) data['cvc'] = this.cvc;
     if (this.token != null) data['token'] = this.token;
-    return data;
-  }
-}
-
-class Extra {
-  Contact shippingContact;
-  Contact billingContact;
-
-  Extra(
-      {this.shippingContact,
-      this.billingContact});
-
-  factory Extra.fromJson(Map<dynamic, dynamic> json) {
-    return Extra(
-      shippingContact: Contact.fromJson(json['shippingContact']),
-      billingContact: Contact.fromJson(json['billingContact']),
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.shippingContact != null) data['shippingContact'] = this.shippingContact.toJson();
-    if (this.billingContact != null) data['billingContact'] = this.billingContact.toJson();
-    return data;
-  }
-}
-
-class Contact {
-  String name;
-  String emailAddress;
-  String phoneNumber;
-  String country;
-  String postalCode;
-  String state;
-  String city;
-  String street;
-  String isoCountryCode;
-  String subAdministrativeArea;
-  String subLocality;
-  String supplementarySubLocality;
-
-  Contact (
-      {this.name,
-      this.emailAddress,
-      this.phoneNumber,
-      this.country,
-      this.postalCode,
-      this.state,
-      this.city,
-      this.street,
-      this.isoCountryCode,
-      this.subAdministrativeArea,
-      this.subLocality,
-      this.supplementarySubLocality
-      });
-
-  factory Contact.fromJson(Map<dynamic, dynamic> json) {
-    print(json);
-    return Contact(
-      name: json['name'] ?? "",
-      emailAddress: json['emailAddress'] ?? "",
-      phoneNumber: json['phoneNumber'] ?? "",
-      country: json['country'] ?? "",
-      postalCode: json['postalCode'] ?? "",
-      state: json['state'] ?? json['administrativeArea'] ?? "",
-      city: json['city'] ?? json['locality'] ?? "",
-      street: json['street'] ?? (json['address1'] != null ? json['address1'] + json['address2'] : ""),
-      isoCountryCode: json['ISOCountryCode'] ?? json['countryCode'] ?? "",
-      subAdministrativeArea: json['subAdministrativeArea'] ?? "",
-      subLocality: json['subLocality'] ?? json['locality'] ?? "",
-      supplementarySubLocality: json['supplementarySubLocality'] ?? "",
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    if (this.name != null) data['name'] = this.name;
-    if (this.emailAddress != null) data['emailAddress'] = this.emailAddress;
-    if (this.phoneNumber != null) data['phoneNumber'] = this.phoneNumber;
-    if (this.country != null) data['country'] = this.country;
-    if (this.postalCode != null) data['postalCode'] = this.postalCode;
-    if (this.state != null) data['state'] = this.state;
-    if (this.city != null) data['city'] = this.city;
-    if (this.street != null) data['street'] = this.street;
-    if (this.isoCountryCode != null) data['ISOCountryCode'] = this.isoCountryCode;
-    if (this.subAdministrativeArea != null) data['subAdministrativeArea'] = this.subAdministrativeArea;
-    if (this.subLocality != null) data['subLocality'] = this.subLocality;
-    if (this.supplementarySubLocality != null) data['supplementarySubLocality'] = this.supplementarySubLocality;
     return data;
   }
 }
